users: Extract user meta from cards
User meta data in cards was duplicated across annotations, highlights,
and bookmarks, some with varying behavior. This change consolidates all
of that into one shared component.
